doshell: clean up signal handler warning.
When restoring the signal handlers, the code saved the old
handler which was just a pointer to SIG_IGN. It did nothing
with that value. This was causing a warning with clang, so
the useless assignment is being removed.

Future work: check the return code for all calls to signal().