fixscreen: initialize wline[]
All of the elements of wline[] were not unconditionally initialized.
clang suggested that maybe there could be an instance were wline[]
could be used in a conditional uninitialized. The function has a
high cyclomatic complexity, so this patch just initializes wline[]
to 0 at the beginning of the function just in case the element
used in the comparison could be used uninitialized.